test(activerecord): merge the mysql schema-creation colliding test file pair
Folds packages/activerecord/src/connection-adapters/mysql/schema-creation.test.ts
into its .trails.test.ts sibling and deletes the plain file. Both are trails
inventions: rubyToConventionTs (scripts/test-compare/compare.ts:135) only maps
cases/connection_adapters/*_test.rb under connection-adapters/, so no future
Rails port lands at either path.

Every describe/it name is preserved verbatim; the shared `leased` /
`mysqlConn` helpers were identical in both copies and are de-duplicated. The
eslint no-explicit-any test exclude row is retargeted at the surviving path.
